excel null values
You can fill all blank cells in all columns of an Excel sheet with "NULL" by using a VBA macro. Here's one way to do it:

Press "Alt + F11" to open the VBA editor.
In the editor, click on "Insert" and then select "Module".
In the new module, paste the following code:

Press "Alt + Q" to return to the worksheet.
Press "Alt + F8" to open the Macro dialog box.
Select the macro named "FillBlanks" and click on "Run".
The macro will fill all blank cells in the entire sheet with "NULL".
Note: The code above will only fill blank cells within the used range of the sheet. If you want to fill all cells, including those that are unused, you will need to change the code to include the entire sheet range.